+++ core/src/main/java/com/andre601/oneversionremake/core/enums/ProtocolVersion.java
@@ -1,0 +1,87 @@
+/*
+ * Copyright 2020 Andre601
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
+ * documentation files (the "Software"), to deal in the Software without restriction, including without limitation
+ * the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software,
+ * and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all copies or substantial
+ * portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+ * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
+ * IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
+ * WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE
+ * OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+package com.andre601.oneversionremake.core.enums;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+public enum ProtocolVersion{
+    MC_1_16_4(754, "1.16.4"),
+    MC_1_16_3(753, "1.16.3"),
+    MC_1_16_2(751, "1.16.2"),
+    MC_1_16_1(736, "1.16.1"),
+    MC_1_16  (735, "1.16"),
+    MC_1_15_2(578, "1.15.2"),
+    MC_1_15_1(575, "1.15.1"),
+    MC_1_15  (573, "1.15"),
+    MC_1_14_4(498, "1.14.4"),
+    MC_1_14_3(490, "1.14.3"),
+    MC_1_14_2(485, "1.14.2"),
+    MC_1_14_1(480, "1.14.1"),
+    MC_1_14  (477, "1.14"),
+    MC_1_13_2(404, "1.13.2"),
+    MC_1_13_1(401, "1.13.1"),
+    MC_1_13  (393, "1.13"),
+    MC_1_12_2(340, "1.12.2"),
+    MC_1_12_1(338, "1.12.1"),
+    MC_1_12  (335, "1.12"),
+    MC_1_11_2(316, "1.11.2"),
+    MC_1_11  (315, "1.11"),
+    MC_1_10_2(210, "1.10.2"),
+    MC_1_9_4 (110, "1.9.4"),
+    MC_1_9_2 (109, "1.9.2"),
+    MC_1_9_1 (108, "1.9.1"),
+    MC_1_9   (107, "1.9"),
+    MC_1_8_9 (47,  "1.8.9"),
+    
+    // In case none of the above versions match
+    UNKNOWN  (0,   "?");
+    
+    private final int protocol;
+    private final String friendlyName;
+    
+    ProtocolVersion(int protocol, String friendlyName){
+        this.protocol = protocol;
+        this.friendlyName = friendlyName;
+    }
+    
+    public static String getFriendlyName(int protocol){
+        for(ProtocolVersion version : values()){
+            if(version.getProtocol() == protocol)
+                return version.getFriendlyName();
+        }
+        
+        return UNKNOWN.getFriendlyName();
+    }
+    
+    public static String getFriendlyNames(List<Integer> protocols){
+        return protocols.stream()
+                .sorted()
+                .map(ProtocolVersion::getFriendlyName)
+                .collect(Collectors.joining(", "));
+    }
+    
+    private int getProtocol(){
+        return protocol;
+    }
+    
+    private String getFriendlyName(){
+        return friendlyName;
+    }
+}
