--- src/main/resources/config.yml
@@ -1,77 +1,0 @@
-##########################################################
-#       ____ _    ______                                 #
-#      / __ \ |  / / __ \            OneVersionRemake    #
-#     / / / / | / / /_/ /                by Andre_601    #
-#    / /_/ /| |/ / _, _/                                 #
-#    \____/ |___/_/ |_|                                  #
-#                                                        #
-#    GitHub:                                             #
-#        https://github.com/Andre601/OneVersionRemake    #
-#                                                        #
-#    Spigot:                                             #
-#        https://spigotmc.org/resources/71727/           #
-##########################################################
-#
-# Setting for what protocol version should be used and how OVR should handle joining players.
-#
-Protocol:
-  #
-  # The protocol versions that should be used.
-  # If the player doesn't have any of these versions will they be not allowed to join the network.
-  #
-  # A list of all supported protocols for the plugin can be found here:
-  #   https://minecraft.gamepedia.com/Protocol_version#Java_Edition_2
-  #
-  Versions: []
-  #
-  # Should OneVersionRemake log any denied Login attempts in the console?
-  # The message will contain the version the player tried to join with.
-  #
-  LogDenial: true
-
-#
-# The various messages you can change.
-#
-Messages:
-  #
-  # This message would be shown where usually the current player count is displayed.
-  # Color and formatting codes are supported.
-  #
-  # Set this to an empty String ('') to disable.
-  #
-  # Placeholder:
-  #   {version}     ->  Displays the version a player needs to join
-  #                     The displayed version is set through the above Protocol number (e.g. 573 becomes 1.15)
-  #   {userVersion} ->  Displays the version which the player is using.
-  #
-  # PlayerCount: '' # Uncomment this and comment out the below option, to disable this.
-  PlayerCount: '&cMinecraft {version}'
-  #
-  # Message shown when the player gets kicked while using an unsupported version.
-  # Color and formatting codes are supported.
-  #
-  # Placeholder:
-  #   {version}     ->  Displays the versions a player needs to join
-  #                     The displayed versions are set through the above Protocol numbers (e.g. 573 becomes 1.15)
-  #   {userVersion} ->  Displays the version which the player is using.
-  #
-  Kick:
-  - '&cYou are using an unsupported version of Minecraft ({userVersion})!'
-  - '&cThis server supports the following Versions: {version}'
-  - '&cPlease change your Version and try again.'
-  #
-  # This text is shown when the cursor of the player hovers over the player count, which normally displays the current
-  # online players.
-  # Color and formatting codes are supported.
-  #
-  # Change this to "Hover: []" to disable
-  #
-  # Placeholder:
-  #   {version}     ->  Displays the version a player needs to join
-  #                     The displayed version is set through the above Protocol number (e.g. 573 becomes 1.15)
-  #   {userVersion} ->  Displays the version which the player is using.
-  #
-  # Hover: [] # uncomment this and comment out the below lines to disable this option.
-  Hover:
-  - '&cYou are using an unsupported version of Minecraft ({userVersion})!'
-  - '&cPlease change your version to {version}.'
